# docs/master_swagger.yaml
basePath: /api/
definitions:
...
  curd.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  curd.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  curd.ReplyData-model_Plugin:
    properties:
      data:
        $ref: '#/definitions/types.Plugin'
      error:
        type: string
    type: object
  curd.ReplyData-model_Product:
    properties:
      data:
        $ref: '#/definitions/types.Product'
      error:
        type: string
    type: object
  curd.ReplyData-model_Role:
    properties:
      data:
        $ref: '#/definitions/types.Role'
      error:
        type: string
    type: object
...
  curd.ReplyList-model_Plugin:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Plugin'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-model_Product:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Product'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-model_Role:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Role'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.ModAggregator:
    properties:
      assign:
        description: 赋值
        type: string
      crontab:
        description: 定时计划
        type: string
      expression:
        description: 表达式
        type: string
      type:
        description: 聚合算法 inc dec avg count min max sum last first
        type: string
    type: object
  types.ModArgument:
    properties:
      desc:
        type: string
      label:
        type: string
      name:
        type: string
      type:
        type: string
      unit:
        type: string
    type: object
  types.ModEvent:
    properties:
      desc:
        type: string
      label:
        type: string
      name:
        type: string
      output:
        items:
          $ref: '#/definitions/types.ModArgument'
        type: array
      type:
        description: info alert error //Level  uint8         `json:"level"`
        type: string
    type: object
  types.ModFunction:
    properties:
      async:
        type: boolean
      desc:
        type: string
      input:
        items:
          $ref: '#/definitions/types.ModArgument'
        type: array
      label:
        type: string
      name:
        type: string
      output:
        items:
          $ref: '#/definitions/types.ModArgument'
        type: array
    type: object
  types.ModParameter:
    properties:
      default:
        type: number
      label:
        type: string
      max:
        type: number
      min:
        type: number
      name:
        type: string
    type: object
  types.ModProperty:
    properties:
      desc:
        type: string
      label:
        type: string
      mode:
        description: r w rw
        type: string
      name:
        type: string
      store:
        description: save diff
        type: string
      type:
        description: int float ....
        type: string
      unit:
        type: string
    type: object
  types.ModValidator:
    properties:
      again:
        description: 再次提醒间隔s
        type: integer
      delay:
        description: 延迟时间s
        type: integer
      expression:
        type: string
      level:
        type: integer
      template:
        type: string
      title:
        type: string
      total:
        description: 总提醒次数
        type: integer
      type:
        type: string
    type: object
...
  types.Plugin:
    properties:
      command:
        type: string
      created:
        type: string
      dependencies:
        additionalProperties:
          type: string
        type: object
      disabled:
        type: boolean
      external:
        type: boolean
      id:
        type: string
      name:
        type: string
      password:
        type: string
      running:
        type: boolean
      username:
        type: string
      version:
        type: string
    type: object
  types.Product:
    properties:
      aggregators:
        items:
          $ref: '#/definitions/types.ModAggregator'
        type: array
      created:
        type: string
      desc:
        type: string
      events:
        items:
          $ref: '#/definitions/types.ModEvent'
        type: array
      functions:
        items:
          $ref: '#/definitions/types.ModFunction'
        type: array
      id:
        type: string
      name:
        type: string
      parameters:
        items:
          $ref: '#/definitions/types.ModParameter'
        type: array
      properties:
        items:
          $ref: '#/definitions/types.ModProperty'
        type: array
      validators:
        items:
          $ref: '#/definitions/types.ModValidator'
        type: array
      version:
        type: string
    type: object
  types.Role:
    properties:
      created:
        type: string
      id:
        type: string
      name:
        type: string
      privileges:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师接口文档
  version: 3.2 版本
paths:
...
  /plugin/{id}:
    post:
      consumes:
      - application/json
      description: 修改插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      - description: 插件信息
        in: body
        name: plugin
        required: true
        schema:
          $ref: '#/definitions/types.Plugin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 修改插件
      tags:
      - plugin
  /plugin/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 删除插件
      tags:
      - plugin
  /plugin/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 禁用插件
      tags:
      - plugin
  /plugin/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 启用插件
      tags:
      - plugin
  /plugin/{id}/start:
    get:
      consumes:
      - application/json
      description: 启动插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 启动插件
      tags:
      - plugin
  /plugin/{id}/stop:
    get:
      consumes:
      - application/json
      description: 停止插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 停止插件
      tags:
      - plugin
  /plugin/create:
    post:
      consumes:
      - application/json
      description: 创建插件
      parameters:
      - description: 插件信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Plugin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Plugin'
      summary: 创建插件
      tags:
      - plugin
...
  /plugin/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入插件
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入插件
      tags:
      - plugin
  /plugin/list:
    get:
      consumes:
      - application/json
      description: 查询插件
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Plugin'
      summary: 查询插件
      tags:
      - plugin
  /plugin/search:
    post:
      consumes:
      - application/json
      description: 查询插件
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Plugin'
      summary: 查询插件
      tags:
      - plugin
...
  /product/{id}:
    get:
      consumes:
      - application/json
      description: 获取产品
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Product'
      summary: 获取产品
      tags:
      - product
    post:
      consumes:
      - application/json
      description: 修改产品
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 产品信息
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/types.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Product'
      summary: 修改产品
      tags:
      - product
  /product/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除产品
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Product'
      summary: 删除产品
      tags:
      - product
  /product/count:
    post:
      consumes:
      - application/json
      description: 查询产品数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询产品数量
      tags:
      - product
  /product/create:
    post:
      consumes:
      - application/json
      description: 创建产品
      parameters:
      - description: 产品信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Product'
      summary: 创建产品
      tags:
      - product
...
  /product/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入产品
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入产品
      tags:
      - product
  /product/list:
    get:
      consumes:
      - application/json
      description: 查询产品
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Product'
      summary: 查询产品
      tags:
      - product
  /product/search:
    post:
      consumes:
      - application/json
      description: 查询产品
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Product'
      summary: 查询产品
      tags:
      - product
  /role/{id}:
    get:
      consumes:
      - application/json
      description: 获取角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Role'
      summary: 获取角色
      tags:
      - role
    post:
      consumes:
      - application/json
      description: 修改角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/types.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Role'
      summary: 修改角色
      tags:
      - role
  /role/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Role'
      summary: 删除角色
      tags:
      - role
  /role/count:
    post:
      consumes:
      - application/json
      description: 查询角色数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询角色数量
      tags:
      - role
  /role/create:
    post:
      consumes:
      - application/json
      description: 创建角色
      parameters:
      - description: 角色信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Role'
      summary: 创建角色
      tags:
      - role
...
  /role/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入角色
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入角色
      tags:
      - role
  /role/list:
    get:
      consumes:
      - application/json
      description: 查询角色
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Role'
      summary: 查询角色
      tags:
      - role
  /role/search:
    post:
      consumes:
      - application/json
      description: 这里写描述 get roles
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Role'
      summary: 查询角色
      tags:
      - role
...
swagger: "2.0"